Add files via upload
car colour detection Model: Description: In this task, you will develop a machine learning model to predict the colour of cars in traffic and count the number of cars at a traffic signal. The model should show a red rectangle for blue cars, and blue rectangles for other colour cars. Additionally, if there are people at the traffic signal, the model should show the number of people present. Guidelines: You should have a proper GUI with a preview of input images.